requirements.txt
pipecat-ai>=0.0.1
